Update README.md
 ;.

The purpose of this project is to analyze machine learning algorithms by looking at red and white wine datasets. The data includes 12 input features: 
•	Fixed Acidity
Titratable acidity sometimes referred to as fixed acidity, is a measurement of the total concentration of titratable acids and free hydrogen ions present in your wine. 
•	Volatile Acidity
Volatile acidity is mostly caused by bacteria in the wine creating acetic acid — the acid that gives vinegar its characteristic flavor and aroma . 
•	Citric Acid
Citric acid is generally found in very small quantities in wine grapes. It acts as a preservative and is added to wines to increase acidity, complement a specific flavor or prevent ferric hazes. It can be added to finished wines to increase acidity and give a “fresh” flavor. Excess addition, however, can ruin the taste.
•	Residual Sugars
RS for short refers to any natural grape sugars that are leftover after fermentation ceases (whether on purpose or not). The juice of wine grapes starts intensely sweet, and fermentation

 Chloride
The amount of chlorides present in wine is usually an indicator of its “saltiness.” This is usually influenced by the territory where the wine grapes grew, cultivation methods, and also the grape type. Too much saltiness is considered undesirable. The right proportion can make the wine more savory.
•	Sulfur Dioxide levels
Sulfur dioxide exists in wine in free and bound forms. It’s the most common preservative used, usually added by winemakers to protect the wine from negative effects of exposure to air and oxygen. Wines with added Sulphur dioxide contents usually have “Contains Sulphites” on their labels. 
•	Free sulfur dioxide  
Prevents microbial growth and oxidation of the wine.
•	Total sulfur dioxide  
Mostly undetectable in wine.
•	Density
Also known as specific gravity, it can be used to measure the alcohol concentration in wines. During fermentation, the sugar in the juice is converted into ethanol with carbon dioxide as waste gas.
•	pH
pH stands for the power of hydrogen, which is a measurement of the hydrogen ion concentration in the solution. 
•	Sulphates
Sulfates are salts of sulfuric acid.
•	Alcohol  
the percent alcohol content of the wine.

We will output the quality of each wine with a score ranging from 1 – 9, base on Tableau to analyze the most important trend, we would use Linear Regression, Random Forest Regression, Elastic Net Regression, and Neural Net Regression. Finally, we will fine-tune the best model for maximum optimization.